requirements.txt
langchain>=0.0.300
chromadb>=0.3.28
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
tqdm>=4.65.0
pydantic>=1.10.7
pytest>=7.2.0
